requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic[email]==2.9.0
pydantic-settings==2.5.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
stripe==10.0.0
httpx==0.27.0
openai==1.50.0
anthropic==0.34.0
duckduckgo-search==6.2.0
celery[redis]==5.4.0
redis==5.1.0
python-multipart==0.0.9
loguru==0.7.2
